feat：add messageTask modular (#7)
* feat：add channel

* refactor：add set and get of channel parameters in aggregation root

* style：adjust code format

* refactor：add EntityDto

* refactor: adjust channel and add notificationTemplate code fist

* refactor：add some infrastructure

* refactor：Add ChannelCaller

* refactor：remove .template.config folder

* feat(Channel Blazor)：Add channel blazor

* style：Code formatting

* style：Code formatting

* refactor：AuditEntity user ID cannot be empty

* refactor：delete ctor

* refactor：Add handle clear to the search box

* refactor：Add date formatting

* refactor(Channel)：blazor Inject Mapper

* refactor：Add Loading

* refactor：Blazro Localization JSON key Canonical name

* refactor：Manually update extended properties because onchange fails

* refactor(channel)：queryParam add filter

* refactor：formatting code,add formatting code

* style：code format for ValueChanged

* refactor：Replace automapper with mapster

* refactor：Notification -> Message,remove federated primary key

* refactor：File name Notification -> Message

* refactor：Normalized code

* refactor：ExtraProperty use ConcurrentDictionary

* feat：Add MessageTemplate

* refactor：Update Masa.Utils.Expressions

* refactor：code with pascal

* refactor：DDD ->Ddd

* refactor:clear

* refactor:add channel

* refactor：clean

* refactor:add messageTemplate

* refactor：Dealing with consolidation issues

* feat(messageTemplate)： Add Sms Email WebsiteMessage message template

* refactor：MessageTemplateCreateUpdateDto No mapping MessageTemplate

* refactor：Refactoring repository queries

* refactor：remove messageTemplateExample and add dateRange

* feat：Add Infrastructure.Sms.Aliyun

* refactor：QuerySmsTemplate

* refactor(MessageTemplate)：GetAsync use GetWithDetailQueryAsync

* refactor：Query supports user-defined sorting

* refactor：Infrastructure.Sms

* refactor：Remove local aliyunsms parameter

* style：formatting code

* refactor：Some finishing

* refactor：MessageTemplate add templateType filter

* refactor：AddAliyunSms ServiceCollectionExtensions

* refactor：Update MessageTemplateAuditStatusChangedEventHandler

* style：Adjust format

* refactor： Add CollectionExtensions

* refactor：MessageTemplateItem constructor public changed to protected

* refactor：Property remove virtual

* refactor：Restructure Infrastructure.Sms.Aliyun

* Delete src/Infrastructure/MASA.MC.Infrastructure.Common/Helper directory

* Delete src/Services/MASA.MC.Service directory

* refactor：Process folder capitalization

* refactor：Modify switch syntax

* refactor：global using

* refactor：modify messageTemplate aggregateRoot

* refactor：add smsChannelChangedDomainEvent

* refactor：add i18n json

* refactor：add pagination component

* refactor：restructure caller and update package

* refactor： add smsTemplate

* refactor： Add Synchronize Alibaba cloud SMS template,To be optimized

* refactor：Add SMS template manual synchronization

* refactor：CallerProvider replace dictionary with object

* fix：Fix the problem that the deletion operation cannot catch exception tips

* Create README.md

* refactor：Add verification

* If mstepper is used with mform, the error handler component will not catch exceptions

* feat：add message task basic code

* Process shared entity appchannel

* Handling appchannel migration

* refactor：refactoring aggregateRoot

* refactor： add messageTask domainService,caller and execution

* refactor：refactoring smstemplatesynchro

* refactor：add some pages

* refactor：Some minor modifications

* refactor：Some minor modifications

* add TemplateMessageCreate

* style：i18n json remove seperator

* style：Usually use . as a separator.

* temp

* feature/messageTemplate

* add MessageTaskListModal and TemplateMessageCreateModal

* feat：Add MessageTask Create,Edit,List

* feat： add send ordinary message and other

* temp,refactoring to be optimized

* refactor： Refactoring message task

* feat：add messageTaskDetail

* feat：Create and edit message tasks and split events by type

* fix：The task of processing and modifying messages is to change the template, resulting in abnormal template variables

* refactor(messageTask)：Refactoring aggregation roots and domain services

* feat：Add WithdrawnHistory，Enabled，Disable

* from main

* feat：Add SendTestMessageTask

* feat：Add GenerateImportTemplate

* refactor：messageTaskQueryHandler remove messageTemplateRepository

* refactor：add convertToType

* refactor：add docker

* refactor：Process code review

* refactor：add nav icon

* refactor：remove redundant menu

* refactor(messageTemplate)：When the template parameters are changed, they are changed to the template title and content

* refactor：Add ISoftDelete

* refactor：modify style

* refactor：Refactor SmsTemplateSynchroEventHandler

* style:add license header

* style：modify style

* feat：Add Infrastructure.Email

* feat：add messageInfos and refactor receivers

* style：handling naming conventions

* style：style adjustment

* style：remove the annotated code

* refactor：temp

* style：adjust some code specifications

* style：adjust some code specifications

* style：adjust some code specifications

* feat：Add SendEmailMessage

* feat(MessageRecord)：Add MessageRecord Api

* feat：Add MessageRecordManagement

* feat：add MessageRecord api and MessageRecordDetailModal

* refactor：Check whether the SHORT message is sent successfully

* refactor：modify profile

* refactor：Use messagetaskhistory as a separate aggregation root

* style：Modify parameter naming

* refactor：modify MessageSendingRules.razor

* style：remove using

* refactor：change pageSize

* style： update en-US.json

* refactor：Minor adjustment

+++ src/Services/Masa.Mc.Service/Application/Channels/ChannelCommandHandler.cs
@@ -1,15 +1,18 @@
-﻿namespace Masa.Mc.Service.Admin.Application.Channels;
+﻿// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
+namespace Masa.Mc.Service.Admin.Application.Channels;
 
 public class ChannelCommandHandler
 {
     private readonly IChannelRepository _repository;
-    private readonly IIntegrationEventBus _integrationEventBus;
+    private readonly IMessageTaskRepository _messageTaskRepository;
     private readonly ChannelDomainService _domainService;
 
-    public ChannelCommandHandler(IChannelRepository repository, IIntegrationEventBus integrationEventBus, ChannelDomainService domainService)
+    public ChannelCommandHandler(IChannelRepository repository, IMessageTaskRepository messageTaskRepository, ChannelDomainService domainService)
     {
         _repository = repository;
-        _integrationEventBus = integrationEventBus;
+        _messageTaskRepository = messageTaskRepository;
         _domainService = domainService;
     }
 
@@ -40,6 +43,10 @@ public class ChannelCommandHandler
         var entity = await _repository.FindAsync(x => x.Id == createCommand.ChannelId);
         if (entity == null)
             throw new UserFriendlyException("channel not found");
+        if (await _messageTaskRepository.FindAsync(x => x.ChannelId == createCommand.ChannelId, false) != null)
+        {
+            throw new UserFriendlyException("There are message tasks in use and cannot be deleted");
+        }
         await _domainService.DeleteAsync(entity);
     }
 }
