+++ src/Services/Masa.Mc.Service/Application/MessageTemplates/MessageTemplateCommandHandler.cs
@@ -1,15 +1,18 @@
-﻿namespace Masa.Mc.Service.Admin.Application.MessageTemplates;
+﻿// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
+namespace Masa.Mc.Service.Admin.Application.MessageTemplates;
 
 public class MessageTemplateCommandHandler
 {
     private readonly IMessageTemplateRepository _repository;
-    private readonly IIntegrationEventBus _integrationEventBus;
+    private readonly IMessageTaskRepository _messageTaskRepository;
     private readonly MessageTemplateDomainService _domainService;
 
-    public MessageTemplateCommandHandler(IMessageTemplateRepository repository, IIntegrationEventBus integrationEventBus, MessageTemplateDomainService domainService)
+    public MessageTemplateCommandHandler(IMessageTemplateRepository repository, IMessageTaskRepository messageTaskRepository, MessageTemplateDomainService domainService)
     {
         _repository = repository;
-        _integrationEventBus = integrationEventBus;
+        _messageTaskRepository = messageTaskRepository;
         _domainService = domainService;
     }
 
@@ -48,6 +51,10 @@ public class MessageTemplateCommandHandler
         var entity = await _repository.FindAsync(x => x.Id == createCommand.MessageTemplateId);
         if (entity == null)
             throw new UserFriendlyException("messageTemplate not found");
+        if (await _messageTaskRepository.FindAsync(x => x.EntityType == MessageEntityTypes.Template && x.EntityId == createCommand.MessageTemplateId, false) != null)
+        {
+            throw new UserFriendlyException("There are message tasks in use and cannot be deleted");
+        }
         await _domainService.DeleteAsync(entity);
     }
 }
